Refresh CI, update the definition of custom configuration
Goign to move from having people define a variable named `schemas` to a
function named `get_config` that returns a bunch of different things for
whylogs/langkit.

Also switch to a single file layout so there's no ambiguity about how to
import things. There's no logic in this package so the risk of the file
getting too big is minimal for the time being.
