+++ Makefile
@@ -1,10 +1,12 @@
-
 project_name := whylogs_container_types
 src := $(shell find $(project_name)/ -name "*.py" -type f)
 
-.PHONY: all
+.PHONY: dist build publish lint test lint-fix format format-fix fix help clean
 
-all: build
+dist: build
+
+clean: ## Clean up build artifacts
+	rm -rf dist
 
 build: $(src)
 	poetry build
@@ -18,13 +20,21 @@ lint: ## Check for type issues with pyright
 test: ## Run unit tests
 	poetry run pytest
 
+lint: ## Check for type issues with pyright
+	@{ echo "Running pyright\n"; poetry run pyright; PYRIGHT_EXIT_CODE=$$?; } ; \
+	{ echo "\nRunning ruff check\n"; poetry run ruff check; RUFF_EXIT_CODE=$$?; } ; \
+	exit $$(($$PYRIGHT_EXIT_CODE + $$RUFF_EXIT_CODE))
+
+lint-fix:
+	poetry run ruff check --fix
+
 format: ## Check for formatting issues
-	poetry run black --check $(project_name)
-	poetry run autoflake --check --in-place --remove-unused-variables $(src)
+	poetry run ruff format --check
 
 format-fix: ## Fix formatting issues
-	poetry run black $(project_name)
-	poetry run autoflake --in-place --remove-unused-variables $(src)
+	poetry run ruff format
+
+fix: lint-fix format-fix ## Fix all linting and formatting issues
 
 help: ## Show this help message.
 	@echo 'usage: make [target] ...'
