+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "whylogs-container-types"
-version = "0.4.0"
+version = "0.4.4"
 description = ""
 authors = ["Anthony Naddeo <anthony@whylabs.ai>"]
 license = "MIT"
@@ -10,28 +10,46 @@ packages = [{ include = "whylogs_container_types/**/*.py" }]
 [tool.poetry.dependencies]
 python = "^3.8"
 whylogs = "*"
+langkit = {url = "https://whypy.s3.us-west-2.amazonaws.com/langkit-0.0.31-py3-none-any.whl"}
 
 
 [tool.poetry.group.dev.dependencies]
 pyright = "^1.1.337"
-black = "^23.11.0"
-autoflake = "^2.2.1"
+ruff = "^0.1.13"
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
-[tool.black]
-line-length = 140
+[project.urls]
+Homepage = "https://github.com/whylabs/whylogs-container-types"
 
 [tool.pyright]
-include = ["whylogs_container"]
+include = ["whylogs_container_types", "integ"]  # TODO add tests to this also once we fix their types
 typeCheckingMode = "strict"
 
 reportMissingTypeStubs = false
 reportMissingParameterType = false
 reportMissingTypeArgumet = false
 
+[tool.ruff]
+line-length = 140
+indent-width = 4
+include = ["./whylogs_container_types/**/*.py", "./tests/**/*.py", "./integ/**/*.py"]
+select = ["E", "F", "I", "W"]
 
-[project.urls]
-Homepage = "https://github.com/whylabs/whylogs-container-types"
+[tool.ruff.isort]
+known-first-party = ["whylogs", "langkit"]
+
+
+[tool.ruff.lint]
+fixable = ["ALL"]
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
+
+
